third_party/WebKit/LayoutTests/fullscreen/full-screen-twice.html

<body>
<script src="full-screen-test.js"></script>
<span></span>
<script>
    var span = document.getElementsByTagName("span")[0];

    var spanEnteredFullScreen = function() {
        testExpected("document.webkitCurrentFullScreenElement", span);
        waitForEventAndEnd(document, "webkitfullscreenerror");
        runWithKeyDown(function(){document.documentElement.webkitRequestFullScreen()});
    };

    waitForEvent(document, "webkitfullscreenchange", spanEnteredFullScreen);
    runWithKeyDown(function(){span.webkitRequestFullScreen()});
</script>
