third_party/WebKit/LayoutTests/fullscreen/full-screen-twice-expected.txt

EVENT(webkitfullscreenchange)
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLSpanElement]') OK
EVENT(webkitfullscreenerror)
END OF TEST
